+++ compass/extraction/date.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from compass.llm import LLMUsageCategory
+from compass.utilities.enums import LLMUsageCategory
 
 
 logger = logging.getLogger(__name__)
