+++ compass/validation/content.py
@@ -8,7 +8,7 @@ import asyncio
 import logging
 from abc import ABC, abstractmethod
 
-from compass.llm import LLMUsageCategory
+from compass.utilities.enums import LLMUsageCategory
 
 
 logger = logging.getLogger(__name__)
